requirements.txt
accelerate==0.23.0
aiocron==1.8
aiofiles==23.2.1
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anthropic==0.8.1
anyio==3.7.1
anytree==2.12.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==5.3.3
camel-converter==3.1.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
controlnet-aux==0.0.7
croniter==2.0.2
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.4
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.21.4
dill==0.3.7
distro==1.9.0
Django==5.0.2
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.7.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
faiss-cpu==1.7.4
fastapi==0.103.2
fastjsonschema==2.18.1
ffmpy==0.3.1
filelock==3.12.4
fire==0.4.0
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
gitignore_parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gradio==3.47.1
gradio_client==0.6.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.33.0
hyperframe==6.0.1
idna==3.4
imageio==2.31.5
imap-tools==1.5.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.27.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-path==0.3.2
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter-contrib-core==0.4.2
jupyter-events==0.7.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==7.4.9
jupyter_core==5.4.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lancedb==0.4.0
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.23
langsmith==0.0.87
lazy-object-proxy==1.10.0
lazy_loader==0.3
libcst==1.0.1
librosa==0.10.1
llvmlite==0.41.1
loguru==0.6.0
lxml==5.1.0
Markdown==3.8.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
meilisearch==0.21.0
metagpt==0.6.6
mistune==3.0.2
modelscope==1.27.0
more-itertools==10.2.0
motor==3.3.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==6.5.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.3
open-clip-torch==2.22.0
openai==1.6.0
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
openpyxl==3.1.2
orjson==3.9.9
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parse==1.20.1
parso==0.8.3
pathable==0.4.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
playwright==1.41.2
pooch==1.7.0
portalocker==2.8.2
prance==23.6.21.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
py==1.11.0
pyarrow==13.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
pyee==11.0.1
Pygments==2.16.1
PyJWT==2.8.0
pylance==0.9.0
pymongo==4.6.2
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==24.0.1
qdrant-client==1.7.0
ratelimiter==1.2.0.post0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
semantic-kernel==0.4.3.dev0
semantic-version==2.10.0
semver==3.0.2
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==2.0.27
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
ta==0.10.2
tenacity==8.2.2
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tiktoken==0.5.2
timm==0.9.7
tinycss2==1.2.1
tokenizers==0.21.1
toolz==0.12.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.3.3
tqdm==4.65.0
traitlets==5.11.2
transformers==4.52.4
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.23.2
wandb==0.15.12
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
win32-setctime==1.1.0
wrapt==1.15.0
xformers==0.0.22.post4
xxhash==3.4.1
yarl==1.9.2
zhipuai==1.0.7
zipp==3.17.0
